+++ app/routes/__init__.py
@@ -1,0 +1,13 @@
+from .inbox_router import inbox_route
+from .conversation_router import conversation_route
+from .message_router import message_route
+from .contact_router import contact_route
+from .dashboard_router import dashboard_route
+
+__all__ = [
+    "inbox_route",
+    "conversation_route",
+    "message_route",
+    "contact_route",
+    "dashboard_route",
+]
